REFACTOR: Decouple ArribaQueryServices from ArribaQueryApplication
